ready for usa boeing test

+++ recipes-example/discrete-in/files/discrete_in.h
@@ -14,9 +14,9 @@
  * @brief SPI를 통해 SENSE<15:0> 레지스터 값을 읽는 함수
  * @return 레지스터 값
  */
-uint32_t GetDiscreteState(void);
-uint32_t GetDiscreteState15to8(void);
-uint32_t GetDiscreteState7to0(void);
+uint16_t GetDiscreteState(void);
+uint8_t GetDiscreteState15to8(void);
+uint8_t GetDiscreteState7to0(void);
 uint8_t ReadProgramSenseBanks(void);
 void WriteProgramSenseBanks(uint8_t bank_settings);
